+++ secretflow_serving/ops/op_kernel.h
@@ -76,7 +76,8 @@ class OpKernel {
   void Compute(ComputeContext* ctx) {
     int64_t rows = ctx->inputs.front().front()->num_rows();
     SERVING_ENFORCE_EQ(ctx->inputs.size(), input_schema_list_.size(),
-                       "schema size be equal to input edges");
+                       "node: {} schema size be equal to input edges",
+                       opts_.node_def.name());
 
     for (size_t edge_index = 0; edge_index != ctx->inputs.size();
          ++edge_index) {
@@ -85,7 +86,8 @@ class OpKernel {
            ++party_index) {
         auto& input_table = edge_inputs[party_index];
         SERVING_ENFORCE_EQ(rows, input_table->num_rows(),
-                           "rows of all inputs tables should be equal");
+                           "node: {} rows of all inputs tables should be equal",
+                           opts_.node_def.name());
 
         if (input_schema_list_[edge_index]->num_fields() > 0 &&
             !input_table->schema()->Equals(input_schema_list_[edge_index])) {
@@ -108,11 +110,15 @@ class OpKernel {
 
     SERVING_ENFORCE_EQ(rows, ctx->output->num_rows(),
                        "rows of input and output be equal");
-    SERVING_ENFORCE(ctx->output->schema()->Equals(output_schema_),
-                    errors::ErrorCode::LOGIC_ERROR,
-                    "schema of output ({}) should match output_schema ({})",
-                    ctx->output->schema()->ToString(),
-                    output_schema_->ToString());
+    if (output_schema_->num_fields() > 0) {
+      // only check when output schema is valid
+      SERVING_ENFORCE(
+          ctx->output->schema()->Equals(output_schema_),
+          errors::ErrorCode::LOGIC_ERROR,
+          "node: {} schema of output ({}) should match output_schema ({})",
+          opts_.node_def.name(), ctx->output->schema()->ToString(),
+          output_schema_->ToString());
+    }
   }
 
   virtual void DoCompute(ComputeContext* ctx) = 0;